Exit early from sharded search
